explode suite_test.go into cmd testing packages
pattern: for cli 'foo', have a 'footest.Init' which returns a
'footest.Cmd' which is a cmdtest.Cmd but with CLI-specific helper
methods attached
