+++ integration/ops/downgrade_test.go
@@ -4,20 +4,54 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/concourse/concourse/integration/dctest"
+	"github.com/concourse/concourse/integration/flytest"
 	"github.com/stretchr/testify/require"
 )
 
 func TestDowngrade(t *testing.T) {
 	t.Parallel()
 
-	dc := dockerCompose(t)
+	dc := dctest.Init(t)
 
 	t.Run("deploy dev", func(t *testing.T) {
 		require.NoError(t, dc.Run("up", "-d"))
 	})
 
-	fly := initFly(t, dc)
+	fly := flytest.Init(t, dc)
+	setupUpgradeDowngrade(t, fly)
 
+	latestDC := dctest.Init(t, "overrides/latest.yml")
+
+	latest, err := latestDC.Output("run", "web", "migrate", "--supported-db-version")
+	require.NoError(t, err)
+	latest = strings.TrimRight(latest, "\n")
+
+	t.Run("down migrations", func(t *testing.T) {
+		// just to see what it was before
+		err := dc.Run("run", "web", "migrate", "--current-db-version")
+		require.NoError(t, err)
+
+		err = dc.Run("run", "web", "migrate", "--migrate-db-to-version", latest)
+		require.NoError(t, err)
+	})
+
+	t.Run("downgrade to latest", func(t *testing.T) {
+		require.NoError(t, latestDC.Run("up", "-d"))
+	})
+
+	fly = flytest.Init(t, latestDC)
+	verifyUpgradeDowngrade(t, fly)
+
+	t.Run("upgrading after downgrade", func(t *testing.T) {
+		require.NoError(t, dc.Run("up", "-d"))
+	})
+
+	fly = flytest.Init(t, dc)
+	verifyUpgradeDowngrade(t, fly)
+}
+
+func setupUpgradeDowngrade(t *testing.T, fly flytest.Cmd) {
 	t.Run("set up test pipeline", func(t *testing.T) {
 		err := fly.Run("set-pipeline", "-p", "test", "-c", "pipelines/smoke-pipeline.yml", "-n")
 		require.NoError(t, err)
@@ -28,31 +62,14 @@ func TestDowngrade(t *testing.T) {
 		err = fly.Run("trigger-job", "-j", "test/say-hello", "-w")
 		require.NoError(t, err)
 	})
+}
 
-	latestDC := dockerCompose(t, "overrides/latest.yml")
-
-	latest, err := latestDC.Output("run", "web", "migrate", "--supported-db-version")
-	require.NoError(t, err)
-	latest = strings.TrimRight(latest, "\n")
-
-	t.Run("downgrading", func(t *testing.T) {
-		// just to see what it was before
-		err := dc.Run("run", "web", "migrate", "--current-db-version")
-		require.NoError(t, err)
-
-		err = dc.Run("run", "web", "migrate", "--migrate-db-to-version", latest)
-		require.NoError(t, err)
-
-		require.NoError(t, latestDC.Run("up", "-d"))
-	})
-
-	fly = initFly(t, latestDC)
-
+func verifyUpgradeDowngrade(t *testing.T, fly flytest.Cmd) {
 	t.Run("pipeline and build still exists", func(t *testing.T) {
 		err := fly.Run("get-pipeline", "-p", "test")
 		require.NoError(t, err)
 
-		out, err := fly.Output("watch", "-j", "test/say-hello", "-b", "1")
+		out, err := fly.Output("watch", "-j", "test/say-hello")
 		require.NoError(t, err)
 		require.Contains(t, out, "Hello, world!")
 	})
@@ -63,7 +80,7 @@ func TestDowngrade(t *testing.T) {
 	})
 
 	t.Run("can still run checks", func(t *testing.T) {
-		err = fly.Run("check-resource", "-r", "test/mockery")
+		err := fly.Run("check-resource", "-r", "test/mockery")
 		require.NoError(t, err)
 	})
